# src/swagger_data.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/login": {
      "post": {
        "tags": [
          "piazza-raw-controller"
        ],
        "summary": "login",
        "operationId": "loginUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "loginRequestModel",
            "description": "loginRequestModel",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequestModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LoginResultModel"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": [
          "piazza-raw-controller"
        ],
        "summary": "logout",
        "operationId": "logoutUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LogoutResultModel"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/raw/{nid}": {
      "get": {
        "tags": [
          "piazza-raw-controller"
        ],
        "summary": "getAllPosts",
        "operationId": "getAllPostsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nid",
            "in": "path",
            "description": "nid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/raw/{nid}/search": {
      "get": {
        "tags": [
          "piazza-raw-controller"
        ],
        "summary": "getSearch",
        "operationId": "getSearchUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nid",
            "in": "path",
            "description": "nid",
            "required": true,
            "type": "string"
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "keyword",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/raw/{nid}/simplefeed": {
      "get": {
        "tags": [
          "piazza-raw-controller"
        ],
        "summary": "getAllFeedSnippet",
        "operationId": "getAllFeedSnippetUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nid",
            "in": "path",
            "description": "nid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/raw/{nid}/stats": {
      "get": {
        "tags": [
          "piazza-raw-controller"
        ],
        "summary": "getStatistics",
        "operationId": "getStatisticsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nid",
            "in": "path",
            "description": "nid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/raw/{nid}/userprofile": {
      "get": {
        "tags": [
          "piazza-raw-controller"
        ],
        "summary": "getUserProfiles",
        "operationId": "getUserProfilesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nid",
            "in": "path",
            "description": "nid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/raw/{nid}/users": {
      "get": {
        "tags": [
          "piazza-raw-controller"
        ],
        "summary": "getAllUsers",
        "operationId": "getAllUsersUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nid",
            "in": "path",
            "description": "nid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/raw/{nid}/{cid}": {
      "get": {
        "tags": [
          "piazza-raw-controller"
        ],
        "summary": "getContentPerCid",
        "operationId": "getContentPerCidUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nid",
            "in": "path",
            "description": "nid",
            "required": true,
            "type": "string"
          },
          {
            "name": "cid",
            "in": "path",
            "description": "cid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
...
    "LoginRequestModel": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "LogoutResultModel": {
      "type": "object"
    },
    "LoginResultModel": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "timeStamp": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
...
  }
}